HngApi/requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
defusedxml
Django
diff-match-patch
django-cors-headers
django-import-export
djangorestframework
drf-yasg
et-xmlfile
gunicorn
idna
inflection
itypes
Jinja2
MarkupPy
MarkupSafe
odfpy
openpyxl
packaging
Pillow
psycopg2
python-decouple
python-dotenv
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
sqlparse
tablib
typing_extensions
tzdata
uritemplate
urllib3
webencodings
xlrd
xlwt
